requirements.txt
django==2.1.2
django-extensions==2.1.3
certifi==2018.8.24
chardet==3.0.4
idna==2.7
urllib3==1.23
djangorestframework==3.8.2
django-filter==2.0.0

#Auth
djoser==1.3.2
djangorestframework-jwt==1.11.0

#AWS
django-storages==1.7.1
boto3==1.9.70

#Machine learning
numpy==1.15.4
pandas==0.23.4
scikit-learn==0.20.2